feat: load patient on internal using RO DB instance
Ref eng-268
